fog with imagination : third batch

+++ rest-servers/games/imagined_units.py
@@ -15,7 +15,7 @@ class ImaginedUnit:
     """ Class for handling a imagined unit """
 
     @staticmethod
-    def list_by_game_id(sql_executor: database.SqlExecutor, game_id: int) -> typing.List[typing.Tuple[int, int, int, int]]:
+    def list_by_game_id(sql_executor: database.SqlExecutor, game_id: int) -> typing.List[typing.Tuple[int, int, int, int, int]]:
         """ class lookup : finds the object in database from fame id """
         units_found = sql_executor.execute("SELECT * FROM imagined_units where game_id = ?", (game_id,), need_result=True)
         if not units_found:
@@ -23,9 +23,9 @@ class ImaginedUnit:
         return units_found
 
     @staticmethod
-    def list_by_game_id_role_num(sql_executor: database.SqlExecutor, game_id: int, role_num: int) -> typing.List[typing.Tuple[int, int, int]]:
+    def list_by_game_id_role_num(sql_executor: database.SqlExecutor, game_id: int, role_id: int) -> typing.List[typing.Tuple[int, int, int, int, int]]:
         """ class lookup : finds the object in database from fame id """
-        units_found = sql_executor.execute("SELECT * FROM imagined_units where game_id = ? and role_num = ?", (game_id, role_num), need_result=True)
+        units_found = sql_executor.execute("SELECT * FROM imagined_units where game_id = ? and role_id = ?", (game_id, role_id), need_result=True)
         if not units_found:
             return []
         return units_found
@@ -35,12 +35,15 @@ class ImaginedUnit:
         """ creation of table from scratch """
 
         sql_executor.execute("DROP TABLE IF EXISTS imagined_units")
-        sql_executor.execute("CREATE TABLE imagined_units (game_id INTEGER, type_num INTEGER, zone_num INTEGER, role_num INTEGER)")
+        sql_executor.execute("CREATE TABLE imagined_units (game_id INTEGER, role_id INTEGER, type_num INTEGER, zone_num INTEGER, role_num INTEGER)")
 
-    def __init__(self, game_id: int, type_num: int, zone_num: int, role_num: int) -> None:
+    def __init__(self, game_id: int, role_id: int, type_num: int, zone_num: int, role_num: int) -> None:
 
         assert isinstance(game_id, int), "game_id must be an int"
         self._game_id = game_id
+
+        assert isinstance(role_num, int), "role_id must be an int"
+        self._role_id = role_id
 
         assert isinstance(type_num, int), "type_num must be an int"
         self._type_num = type_num
@@ -53,14 +56,14 @@ class ImaginedUnit:
 
     def update_database(self, sql_executor: database.SqlExecutor) -> None:
         """ Pushes changes from object to database """
-        sql_executor.execute("INSERT OR REPLACE INTO imagined_units (game_id, type_num, zone_num, role_num) VALUES (?, ?, ?, ?)", (self._game_id, self._type_num, self._zone_num, self._role_num))
+        sql_executor.execute("INSERT OR REPLACE INTO imagined_units (game_id, role_id, type_num, zone_num, role_num) VALUES (?, ?, ?, ?, ?)", (self._game_id, self._role_id, self._type_num, self._zone_num, self._role_num))
 
     def delete_database(self, sql_executor: database.SqlExecutor) -> None:
         """ Removes object from database """
-        sql_executor.execute("DELETE FROM imagined_units WHERE game_id = ? AND zone_num = ?", (self._game_id, self._zone_num))
+        sql_executor.execute("DELETE FROM imagined_units WHERE game_id = ? AND role_id = ? AND zone_num = ?", (self._game_id, self._role_id, self._zone_num))
 
     def __str__(self) -> str:
-        return f"game_id={self._game_id} type_num={self._type_num} zone_num={self._zone_num} role_num={self._role_num}"
+        return f"game_id={self._game_id} role_id={self._game_id} type_num={self._type_num} zone_num={self._zone_num} role_num={self._role_num}"
 
 
 if __name__ == '__main__':
